requirements.txt
slack_sdk
slack_bolt
google-api-python-client
google-auth-oauthlib
python-dotenv